Intergrated PID Loop and Configured Format
Within PIDController.h is a class based system that allows for individual controllers to be created and used. The PID functions within the controller are that of a standard PID calculation, though it does not execute the loops, meaning that has to be done outside the class. Its implementation can be seen in the main.cpp file, where a controller is created for the left, and the right grouping of motors, and applied to each accordingly.