Ammar's PID and Joystick Curve
My rendition of a PID that is written more traditionally and utilizes the same method of conversion from ticks to inches like in Jonathan's PID. It omits the integral part of a PID for the drivetrain.

My program also features a joystick curve function that allows one to change the sensitivity of the controller in controlling the drivetrain.