Multithreading and Beautification of Code
I've included threads that relate to the controller to allow for two functions to run at the same time in a while loop. This was needed because without it, the curveJoystick function would interfere with other parts of the controller, like button presses; it would make button presses lag and run for longer than intended.

I've also began beautifying my code - simplifying where needed and grouping like commands together. I've added a lot of comments and documentation intended to make it easier for anyone to interpret my code and understand why I wrote some thing.