Jonathan PID Debugging
We fixed a bug in Jonathan's PID in which the PID function would cause the drivetrain to drive forever.
This was caused by one line in the code "resetMotorValues();" being placed in the PID function. The PID function runs forever in a loop, meaning the motor values would be reset to position 0 infinitely. This leads to the motors being they are never reaching the target distance and will move forever.

 This was fixed by removing the resetMotorValues() function from the PID function and instead moving it to the autonomous code. This way, the motors still get reset at the beginning of every run.