Reverting Turning PID Implementation
We realized that the implementation of the turning PID into the same PID function resulted in the error fluctuating into outlier values, rather than slowly reducing to (hopefully!) 0. Due to this, we want to revert our changes back to the original PID class and rethink a better way to steer the robot precisely and smoothly. We are looking into Odometry programming currently.