Limit Switch Catapult + Pneumatics
We added functionality for the Catapult, such that it runs forever until it comes into contact with the limit switch in the drivetrain. When it comes into contact with the limit switch, the motor stops - but is still applying power to hold the motor in place - until our driver presses R2 on the Controller, which resumes the catapult motor and sets off the catapult.

We also added user control for the pneumatics that trigger the flaps on the sides. The program sets the pneumatics to false in pre-autonomous, and awaits our driver to press L2, which will toggle our flaps on the sides of the drivetrains when we are ready to push the triballs in.

Misc. changes include setting drivetrain and catapult motor velocity to maximum speed and minor PID changes.

Co-Authored-By: mochi4852 <127371243+mochi4852@users.noreply.github.com>