Distance Sensor Tuning - 10/31 and 11/1
I am testing what is the most ideal distance for the catapult to stop at such that the distance sensor does not glitch and fail to provide a reading, which may cause the catapult to stop functioning.
I am also experimenting with the brakeType modes for the catapult that will work best with the ratchet.