Changes to port of catapultMotor + brakeMode, etc.
Changed the catapultMotor port from PORT10 to PORT9.
Temporarily disabled distance sensor automated braking until ratchet is installed on robot.
Experimenting with brakeTypes on the catapults; brake seems to hold the motor when stopped exactly, but overheats the motors rather quickly
Hold also attempts to stop the motors, but has a lot more twitching before setlling; small movements cause it to readjust

Added a line to make the inertial sensor on the robot calibrate before running the autonomous. This should prevent the endless spinning.
Continuing to tune the turnPID